requirement.txt
python==3.8.17
jupyter==1.0.0
jupyter_client==7.4.9
jupyter-console==6.6.3
jupyter_core==5.3.0
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter-server==1.23.4
jupyter_server_terminals==0.4.4
matplotlib==3.7.1
matplotlib-inline==0.1.6
numpy==1.24.3
openpyxl==3.1.2
pandas==1.5.3
Pillow==9.4.0
pip==23.2.1